backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
yt-dlp==2025.2.19
openai-whisper==20240930
ffmpeg-python==0.2.0
pydantic==2.10.4
python-multipart==0.0.20
